inference_segmentation.py: add support for stac item as direct input aoi.py: move stac class to its own file geoutils.py: convert to absolute path if not url test_stacitem.py: move stac-specific tests from test_aoi.pu
