+++ config/inference/default_binary.yaml
@@ -1,6 +1,7 @@
 # @package _global_
 inference:
   raw_data_csv: tests/inference/inference_segmentation_binary.csv
+  input_stac_item:  # alternatively, use a path or url to stac item directly
   state_dict_path: ${general.save_weights_dir}/
   chunk_size:  # if empty, will be calculated automatically from max_pix_per_mb_gpu
   # Maximum number of pixels each Mb of GPU Ram to allow. E.g. if GPU has 1000 Mb of Ram and this parameter is set to
